requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
six==1.9.0
websocket-client==0.26.0
wsgiref==0.1.2
